[ci][fix] Enable part of ray test to be run on CPU machine (#1372)
