+++ tests/ray_gpu/detached_worker/README.md
