+++ tests/ray_gpu/detached_worker/run.sh
