+++ tests/ray_gpu/test_rvdz.py
