+++ verl/single_controller/base/worker.py
@@ -137,14 +137,11 @@ class Worker(WorkerHelper):
         # construct a meta from environment variable. Note that the import must be inside the class because it is executed remotely
         import os
 
-        ###
-        # [SUPPORT AMD: torch]
         import torch
-        ###
 
         ###
         # [SUPPORT AMD: torch]
-        if "AMD" in torch.cuda.get_device_name():
+        if torch.cuda.is_available() and "AMD" in torch.cuda.get_device_name():
             os.environ["CUDA_VISIBLE_DEVICES"] = os.environ.get("ROCR_VISIBLE_DEVICES")
             os.environ["LOCAL_RANK"] = os.environ.get("RAY_LOCAL_RANK")
         ###
@@ -162,13 +159,8 @@ class Worker(WorkerHelper):
 
         ###
         # [SUPPORT AMD: torch]
-        if "AMD" in torch.cuda.get_device_name():
+        if torch.cuda.is_available() and "AMD" in torch.cuda.get_device_name():
             self.local_rank = int(os.environ["LOCAL_RANK"])
-        ###
-
-        ###
-        # [SUPPORT AMD: torch]
-        if "AMD" in torch.cuda.get_device_name():
             cuda_visible_devices = str(local_rank)
         ###
 
@@ -188,8 +180,7 @@ class Worker(WorkerHelper):
 
         ###
         # [SUPPORT AMD: torch]
-        # torch.cuda.set_device(local_rank)
-        if "AMD" in torch.cuda.get_device_name():
+        if torch.cuda.is_available() and "AMD" in torch.cuda.get_device_name():
             torch.cuda.set_device(int(cuda_visible_devices))
         ###
